fuse/runtests: added 'random scatter' write test
this writes the test file in a randomised order, with randomly sized writes.
also for each 'slice' of the file written, a randomly chosen overlapping
write is also made to the file.  this ensures that the file will be written
in its entirety in a thoroughly random order, with many overlapping writes.

darcs-hash:20081003233436-80e44-7cb31d0decfac8d1cf1d33f66ac80a712d46de85.gz