try to tidy up uri-as-string vs. uri-as-object
Ignore-this: 4280727007c29f5b3e9273a34519893f
I get confused about whether a given argument or return value is a uri-as-string or uri-as-object.  This patch adds a lot of assertions that it is one or the other, and also changes CheckerResults to take objects not strings.
In the future, I hope that we generally use Python objects except when importing into or exporting from the Python interpreter e.g. over the wire, the UI, or a stored file.

darcs-hash:20081219143924-92b7f-c5143e0acb5822496d59b22c4d89a151c0af2214.gz