dirnode: don't check MAC on entries in dirnodes
Ignore-this: efacb56d18259219c910cf5c84b17340
In an ancient version of directories, we needed a MAC on each entry.  In modern times, the entire dirnode comes with a digital signature, so the MAC on each entry is redundant.
With this patch, we no longer check those MACs when reading directories, but we still produce them so that older readers will accept directories that we write.

darcs-hash:20081221233518-92b7f-73eec3757419f77b23db2f52a201d0105172f5a3.gz