cli: make startstop_node wait 20 seconds instead of 5 for a process to go away after we signalled it to go away
Ignore-this: 3b0d47649e32b01ff55a506245c674c6
Because the unit tests on the VirtualZooko buildslave failed when it took 16 seconds for a process to go away.
Perhaps getting notification after only 5 seconds instead of 20 seconds is desirable, and we should change the unit tests and set this back to 5, but I don't know exactly how to change the unit tests.  Perhaps match this particular warning message about the shutdown taking a while and allow the code under test to pass if the only stderr that it emits is this warning.

darcs-hash:20081230072022-92b7f-d9fd0e621f518740ed42dc64dc42618b9c754fb2.gz