immutable: fix the writing of share data size into share file in case the share file is used by a < v1.3.0 storage server
Ignore-this: 7d6025aba05fe8140bb712e71e89f1ba
Brian noticed that the constant was wrong, and in fixing that I noticed that we should be saturating instead of modding.
This code would never matter unless a server downgraded or a share migrated from Tahoe >= v1.3.0 to Tahoe < v1.3.0.  Even in that case, this bug would never matter unless the share size were exactly 4,294,967,296 bytes long.
Brian, for good reason, wanted this to be spelled "2**32" instead of "4294967296", but I couldn't stand to see a couple of more Python bytecodes interpreted in the middle of a core, frequent operation on the server like immutable share creation.

darcs-hash:20090106182404-92b7f-50b2305863645b0f825137345087f99dde3427c9.gz