immutable: redefine the "sharemap" member of the upload results to be a map from shnum to set of serverids
Ignore-this: 10300a2333605bc26c4ee9c7ab7dae10
It used to be a map from shnum to a string saying "placed this share on XYZ server".  The new definition is more in keeping with the "sharemap" object that results from immutable file checking and repair, and it is more useful to the repairer, which is a consumer of immutable upload results.

darcs-hash:20090110174623-92b7f-6f684c343360317b11e685aa32c95e2d33fef9d3.gz