immutable repairer: fix DownUpConnector so that it satisfies short reads the were requested after the last write and before the close
Ignore-this: 2856bb9582d9c38a82fd9aad7fd10b57
This is probably the cause of the very rare "loss of progress" bug.  This is tested by unit tests.  A recent patch changed this to errback instead of losing progress, and now this patch is changing it again to return a short read instead of errbacking.  Returning a short read is what the uploader (in encode.py) is expecting, when it is reading the last block of the ciphertext, which might be shorter than the other blocks.

darcs-hash:20090212230447-92b7f-11ef58cee99e471c69580d675b5bffedc038df29.gz