test_repairer: add (commented-out) test_each_byte, to see exactly what the
Ignore-this: 8e84277d5304752edeff052b97821815
Verifier misses

The results (described in #819) match our expectations: it misses corruption
in unused share fields and in most container fields (which are only visible
to the storage server, not the client). 1265 bytes of a 2753 byte
share (hosting a 56-byte file with an artifically small segment size) are
unused, mostly in the unused tail of the overallocated UEB space (765 bytes),
and the allocated-but-unwritten plaintext_hash_tree (480 bytes).

darcs-hash:20100110203552-66853-3569a732c9517a2e74824b4d2c186f1574c0d4f4.gz