immutable: download from the first servers which provide at least K buckets instead of waiting for all servers to reply
Ignore-this: c855355a40d96827e1d0c469a8d8ab3f
This should put an end to the phenomenon I've been seeing that a single hung server can cause all downloads on a grid to hang.  Also it should speed up all downloads by (a) not-waiting for responses to queries that it doesn't need, and (b) downloading shares from the servers which answered the initial query the fastest.
Also, do not count how many buckets you've gotten when deciding whether the download has enough shares or not -- instead count how many buckets to *unique* shares that you've gotten.  This appears to improve a slightly weird behavior in the current download code in which receiving >= K different buckets all to the same sharenumber would make it think it had enough to download the file when in fact it hadn't.
This patch needs tests before it is actually ready for trunk.

darcs-hash:20100127233417-92b7f-bfc36fbfa56a9b9ad619aaf8521b2e919bf49b2f.gz