immutable: downloader accepts notifications of buckets even if those notifications arrive after he has begun downloading shares.
Ignore-this: 5b09709f27603a3157eba7ba70028955
This can be useful if one of the ones that he has already begun downloading fails. See #287 for discussion. This fixes part of #287 which part was a regression caused by #928, namely this fixes fail-over in case a share is corrupted (or the server returns an error or disconnects). This does not fix the related issue mentioned in #287 if a server hangs and doesn't reply to requests for blocks.

darcs-hash:20100201061610-92b7f-769ae1c9576299dcdb625112c6699d1fe6a10549.gz