setup: remove bundled setuptools_darcs-1.2.8.tar
Ignore-this: 378b1964b513ae7fe22bae2d3478285d
This version of setuptools_darcs had a bug when used on Windows which has been fixed in setuptools_darcs-1.2.9. Hopefully we will not need to bundle a copy of setuptools_darcs-1.2.9 in with Tahoe-LAFS and can instead rely on it to be downloaded from PyPI or bundled in the "tahoe deps" separate tarball.

darcs-hash:20100522015333-92b7f-be3262a4fc858e0a5225d741b37e292b39bd409e.gz