test: increase timeout on a network test because Francois's ARM machine hit that timeout
Ignore-this: 380c345cdcbd196268ca5b65664ac85b
I'm skeptical that the test was proceeding correctly but ran out of time. It seems more likely that it had gotten hung. But if we raise the timeout to an even more extravagant number then we can be even more certain that the test was never going to finish.

darcs-hash:20110317165909-92b7f-341e6d5e987ef43f8cbd1ddf556880db06c94ff6.gz